requirements.txt
pathlib2
patool